Refactor most of the sol-compiler methods into helper functions in utils and make resolver pluggable into them
